third_party/blink/web_tests/svg/dynamic-updates/SVGTextElement-svgdom-textLength-prop-expected.txt

SVG 1.1 dynamic update tests
Stretched text

Tests dynamic updates of the 'textLength' property of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.getAttribute('textLength') is null
PASS lastLength = textElement.getComputedTextLength(); lastLength > 0 && lastLength < 200 is true
PASS lastLength == textElement.textLength.baseVal.value is true
PASS textElement.getAttribute('textLength') is "200"
PASS textElement.textLength.baseVal.value is 200
PASS textElement.getComputedTextLength() is lastLength
PASS successfullyParsed is true

TEST COMPLETE
